TagState: add rename, purge tag methods - make set tags unique - reset meta structure on clear
TagStorage: make listOrphans return actual array

+++ src/APIs/tag/tagStorage.js
@@ -1,7 +1,7 @@
 // @ts-check
 import { JsonStorage } from '../tool/jsonStorage.js';
 import { tagDBFile } from '../tool/appPaths.js';
-import fs from 'fs';
+import { access } from 'fs';
 
 
 /**
@@ -36,7 +36,7 @@ export class TagState extends Map {
    * Tag ID -> Name table.
    * @type {Object<number, string>}
    */
-  #tagId2Name = super.get(TagState.meta.tags);
+  #tagId2Name;
 
   /**
    * Inverted `#tagIdName` table for reverse search. Virtual, not persisted.
@@ -48,13 +48,13 @@ export class TagState extends Map {
    * File Path -> Tag-IDs table.
    * @type {Object<string, number[]>}
    */
-  #file2TagIDs = super.get(TagState.meta.files);
+  #file2TagIDs;
 
   /**
    * Control properties.
    * @type {{nextID: number, orphanIDs: number[]}}
    */
-  #control = super.get(TagState.meta.control);
+  #control;
 
   /**
    * Initialize meta entries.
@@ -62,25 +62,25 @@ export class TagState extends Map {
    */
   constructor(...args) {
     super(...args);
+    this.#generateMetaStructure();
+  }
 
-    if (!this.#file2TagIDs) {
+  /**
+   * Get or create meta entry structure, references.
+   */
+  #generateMetaStructure() {
+    if ( !super.has(TagState.meta.files) )
       super.set(TagState.meta.files, {});
-      this.#file2TagIDs = super.get(TagState.meta.files);
-    }
 
-    if (!this.#tagId2Name) {
+    if ( !super.has(TagState.meta.tags) )
       super.set(TagState.meta.tags, {});
-      this.#tagId2Name = super.get(TagState.meta.tags);
-    }
 
-    if (!this.#control) {
-      super.set(TagState.meta.control, {
-        nextID: 0,
-        orphanIDs: []
-      });
+    if ( !super.has(TagState.meta.control) )
+      super.set(TagState.meta.control, { nextID: 0, orphanIDs: [] });
 
-      this.#control = super.get(TagState.meta.control);
-    }
+    this.#file2TagIDs = super.get(TagState.meta.files);
+    this.#tagId2Name  = super.get(TagState.meta.tags);
+    this.#control     = super.get(TagState.meta.control);
   }
 
   /**
@@ -118,7 +118,7 @@ export class TagState extends Map {
         this.#tagName2Id.set(this.#tagId2Name[id], Number(id) );
     }
 
-    const tagIDs = value.map(name => {
+    const tagIDs = [...new Set(value)].map(name => {
       let tagID = this.#tagName2Id.get(name);
 
       if (tagID == null) {
@@ -130,11 +130,59 @@ export class TagState extends Map {
       return tagID;
     });
 
-    tagIDs.length > 0 ?
-      this.#file2TagIDs[key] = tagIDs :
-      delete this.#file2TagIDs[key];
+    tagIDs.length > 0
+      ? this.#file2TagIDs[key] = tagIDs
+      : delete this.#file2TagIDs[key];
 
     return this;
+  }
+
+  /**
+   * Move entry tags from a path to another.
+   * @param {string} oldPath Old entry key.
+   * @param {string} newPath New entry key.
+   * @return {boolean} Success.
+   */
+  moveEntry(oldPath, newPath) {
+    const tagIDs = this.#file2TagIDs[oldPath];
+
+    if (tagIDs != null) {
+      this.#file2TagIDs[newPath] = tagIDs;
+      delete this.#file2TagIDs[oldPath];
+    }
+
+    return tagIDs != null;
+  }
+
+  /**
+   * Rename all specified tag occurrences.
+   * @param {Map<string, string>} renameMap Old-new tag name key value map.
+   * @returns {number} Entries updated.
+   */
+  renameTags(renameMap) {
+    let entriesUpdated = 0;
+
+    this.forEach((tags, key) => {
+      // unique set of tags
+      const updatedTags = /** @type {Set<string>} */ (new Set());
+      let tagsRenamed = false;
+
+      for (const tag of tags) {
+        const renamedTag = renameMap.get(tag);
+
+        if (renamedTag != null)
+          tagsRenamed = true;
+
+        updatedTags.add(renamedTag ?? tag);
+      }
+
+      if (tagsRenamed) {
+        this.set(key, [...updatedTags]);
+        entriesUpdated++;
+      }
+    });
+
+    return entriesUpdated;
   }
 
   /**
@@ -153,9 +201,35 @@ export class TagState extends Map {
     return existed;
   }
 
+  /**
+   * Delete all occurrences of one or more tags.
+   * @param {string[]} tags Tags to delete.
+   * @returns {number} Entries updated.
+   */
+  deleteTags(tags) {
+    const deletionTagSet = new Set(tags);
+    let entriesUpdated = 0;
+
+    this.forEach((entryTags, key) => {
+      const tagSet = new Set(entryTags);
+      const diffSet = tagSet.difference(deletionTagSet);
+
+      if (tagSet.size > diffSet.size) {
+        this.set(key, [...diffSet]);
+        entriesUpdated++;
+      }
+    });
+
+    return entriesUpdated;
+  }
+
+  /**
+   * Clear state and meta structure.
+   */
   clear() {
     super.clear();
-    this.#tagName2Id.clear(); 
+    this.#generateMetaStructure();
+    this.#tagName2Id.clear();
   }
 
   /**
@@ -228,7 +302,6 @@ export class TagState extends Map {
    * @returns {Set<number>} Deleted tags.
    */
   deleteOrphanedTags() {
-    // console.time('delete orphaned tag');
     const numberIDs = Object.keys(this.#tagId2Name).map(id => Number(id) );
     const tagIDs = new Set(numberIDs);
     const usedIDs = new Set( Object.values(this.#file2TagIDs).flat() );
@@ -240,7 +313,6 @@ export class TagState extends Map {
       this.#control.orphanIDs.push(id);
       delete this.#tagId2Name[id];
     });
-    // console.timeEnd('delete orphaned tag');
 
     return unusedIDs;
   }
@@ -253,13 +325,11 @@ export class TagState extends Map {
    * @returns {Map<string, string[]>}
    */
   solved() {
-    // console.time('deserialize tags');
     const state = /** @type {Map<string, string[]>} */ ( new Map() );
     for (const file in this.#file2TagIDs) {
       const tags = this.#file2TagIDs[file].map(id => this.#tagId2Name[id]);
       state.set(file, tags);
     }
-    // console.timeEnd('deserialize tags');
 
     return state;
   }
@@ -307,8 +377,9 @@ export class TagStorage extends JsonStorage {
 
   /**
    * List database entries whose files are no longer accessible. Optionally clear orphans.
-   * * Changes are persisted.
+   * - Changes are persisted.
    * @param {boolean} [deleteOrphans=false] Either to delete orphaned entries if found.
+   * @returns {Promise<string[]>} Orphaned entries.
    */
   async listOrphans(deleteOrphans = false) {
     const taskPromises = [], orphans = [];
@@ -316,8 +387,8 @@ export class TagStorage extends JsonStorage {
 
     for ( const filepath of state.keys() ) {
       taskPromises.push( new Promise(resolve => {
-        fs.access(filepath, (err) => {
-          if (err) 
+        access(filepath, (err) => {
+          if (err != null) 
             orphans.push(filepath);
 
           resolve(true);
@@ -326,20 +397,14 @@ export class TagStorage extends JsonStorage {
     }
 
     await Promise.all(taskPromises);
-    
-    if (orphans.length < 1) {
-      console.log('\nNo orphan entries to clean.');
-    } else {
-      console.log(`${orphans.length} orphan entries found:`, orphans);
 
-      if (deleteOrphans) {
-        await this.write(db => {
-          for (const key of orphans)
-            db.delete(key);
-        });
-        
-        console.log('Cleaned', orphans.length, 'orphan entries.');
-      }
+    if (deleteOrphans && orphans.length > 0) {
+      await this.write(db => {
+        for (const key of orphans)
+          db.delete(key);
+      });
     }
+
+    return orphans;
   }
 }
